stm32: cleanup ptp
