+++ embassy-stm32/src/eth/mod.rs
@@ -8,25 +8,26 @@ compile_error!("The 'ptp' feature is only supported on STM32 Ethernet MAC v2/v2a
 #[cfg_attr(any(eth_v2, eth_v2a), path = "v2/mod.rs")]
 mod _version;
 mod generic_phy;
-mod packet_state;
+#[cfg(feature = "ptp")]
 mod ptp;
 mod sma;
 
 use core::mem::MaybeUninit;
-use core::ptr::addr_of_mut;
 use core::task::Context;
 
 use embassy_hal_internal::PeripheralType;
-use embassy_net_driver::{Capabilities, HardwareAddress, LinkState, PacketMeta};
+#[cfg(feature = "ptp")]
+use embassy_net_driver::PacketMeta;
+use embassy_net_driver::{Capabilities, HardwareAddress, LinkState};
 use embassy_sync::waitqueue::AtomicWaker;
 
-pub use self::_version::{InterruptHandler, *};
-pub use self::generic_phy::*;
-use self::packet_state::PacketState;
-use self::ptp::PtpTimestampSink;
+pub use crate::eth::_version::{InterruptHandler, *};
+pub use crate::eth::generic_phy::*;
 #[cfg(feature = "ptp")]
-pub use self::ptp::{PtpTimestamp, PtpTimestampStore};
-pub use self::sma::{Instance as SmaInstance, Sma, StationManagement};
+use crate::eth::ptp::{PacketState, PtpTimestampSink};
+#[cfg(feature = "ptp")]
+pub use crate::eth::ptp::{PtpTimestamp, PtpTimestampStore};
+pub use crate::eth::sma::{Instance as SmaInstance, Sma, StationManagement};
 use crate::rcc::RccPeripheral;
 
 #[allow(unused)]
@@ -51,13 +52,17 @@ pub struct PacketQueue<const TX: usize, const RX: usize> {
     rx_desc: [RDes; RX],
     tx_buf: [Packet<TX_BUFFER_SIZE>; TX],
     rx_buf: [Packet<RX_BUFFER_SIZE>; RX],
+    #[cfg(feature = "ptp")]
     packet_state: PacketState<TX, RX>,
 }
 
 impl<const TX: usize, const RX: usize> PacketQueue<TX, RX> {
     /// Create a new packet queue.
     pub const fn new() -> Self {
-        Self::new_inner(PtpTimestampSink::new())
+        Self::new_inner(
+            #[cfg(feature = "ptp")]
+            PtpTimestampSink::new(),
+        )
     }
 
     /// Create a new packet queue with Ethernet PTP packet timestamps.
@@ -74,12 +79,13 @@ impl<const TX: usize, const RX: usize> PacketQueue<TX, RX> {
         Self::new_inner(PtpTimestampSink::from_store(timestamps))
     }
 
-    const fn new_inner(ptp: PtpTimestampSink) -> Self {
+    const fn new_inner(#[cfg(feature = "ptp")] ptp: PtpTimestampSink) -> Self {
         Self {
             tx_desc: [const { TDes::new() }; TX],
             rx_desc: [const { RDes::new() }; RX],
             tx_buf: [Packet([0; TX_BUFFER_SIZE]); TX],
             rx_buf: [Packet([0; RX_BUFFER_SIZE]); RX],
+            #[cfg(feature = "ptp")]
             packet_state: PacketState::new(ptp),
         }
     }
@@ -99,7 +105,8 @@ impl<const TX: usize, const RX: usize> PacketQueue<TX, RX> {
     pub fn init(this: &mut MaybeUninit<Self>) {
         unsafe {
             this.as_mut_ptr().write_bytes(0u8, 1);
-            addr_of_mut!((*this.as_mut_ptr()).packet_state).write(PacketState::new(PtpTimestampSink::new()));
+            #[cfg(feature = "ptp")]
+            (&raw mut (*this.as_mut_ptr()).packet_state).write(PacketState::new(PtpTimestampSink::new()));
         }
     }
 
@@ -115,7 +122,7 @@ impl<const TX: usize, const RX: usize> PacketQueue<TX, RX> {
     ) {
         unsafe {
             this.as_mut_ptr().write_bytes(0u8, 1);
-            addr_of_mut!((*this.as_mut_ptr()).packet_state)
+            (&raw mut (*this.as_mut_ptr()).packet_state)
                 .write(PacketState::new(PtpTimestampSink::from_store(timestamps)));
         }
     }
@@ -135,7 +142,9 @@ impl<'d, T: Instance, P: Phy> embassy_net_driver::Driver for Ethernet<'d, T, P> 
 
     fn receive(&mut self, cx: &mut Context) -> Option<(Self::RxToken<'_>, Self::TxToken<'_>)> {
         WAKER.register(cx.waker());
+        #[cfg(feature = "ptp")]
         self.tx.collect_completed();
+
         if self.rx.available().is_some() && self.tx.available().is_some() {
             Some((RxToken { rx: &mut self.rx }, TxToken { tx: &mut self.tx }))
         } else {
@@ -188,6 +197,7 @@ pub struct RxToken<'a, 'd> {
 }
 
 impl<'a, 'd> embassy_net_driver::RxToken for RxToken<'a, 'd> {
+    #[cfg(feature = "ptp")]
     fn meta(&self) -> PacketMeta {
         self.rx.meta()
     }
@@ -210,6 +220,7 @@ pub struct TxToken<'a, 'd> {
 }
 
 impl<'a, 'd> embassy_net_driver::TxToken for TxToken<'a, 'd> {
+    #[cfg(feature = "ptp")]
     fn set_meta(&mut self, meta: PacketMeta) {
         self.tx.set_meta(meta);
     }
