+++ embassy-stm32/src/eth/v1/rx_desc.rs
@@ -1,10 +1,12 @@
 use core::sync::atomic::{Ordering, compiler_fence, fence};
 
+#[cfg(feature = "ptp")]
 use embassy_net_driver::PacketMeta;
 use stm32_metapac::eth::vals::{Rpd, Rps};
 use vcell::VolatileCell;
 
 use crate::eth::RX_BUFFER_SIZE;
+#[cfg(feature = "ptp")]
 use crate::eth::packet_state::RxPacketStateRing;
 use crate::pac::ETH;
 
@@ -135,6 +137,7 @@ pub enum RunningState {
 pub(crate) struct RDesRing<'a> {
     descriptors: &'a mut [RDes],
     buffers: &'a mut [Packet<RX_BUFFER_SIZE>],
+    #[cfg(feature = "ptp")]
     state: RxPacketStateRing<'a>,
     index: usize,
 }
@@ -143,7 +146,7 @@ impl<'a> RDesRing<'a> {
     pub(crate) fn new(
         descriptors: &'a mut [RDes],
         buffers: &'a mut [Packet<RX_BUFFER_SIZE>],
-        state: RxPacketStateRing<'a>,
+        #[cfg(feature = "ptp")] state: RxPacketStateRing<'a>,
     ) -> Self {
         assert!(descriptors.len() > 1);
         assert!(descriptors.len() == buffers.len());
@@ -161,6 +164,7 @@ impl<'a> RDesRing<'a> {
         Self {
             descriptors,
             buffers,
+            #[cfg(feature = "ptp")]
             state,
             index: 0,
         }
@@ -219,10 +223,12 @@ impl<'a> RDesRing<'a> {
 
         let descriptor = &mut self.descriptors[self.index];
         let len = descriptor.packet_len();
+        #[cfg(feature = "ptp")]
         self.state.capture(self.index, None);
         return Some(&mut self.buffers[self.index].0[..len]);
     }
 
+    #[cfg(feature = "ptp")]
     pub(crate) fn meta(&self) -> PacketMeta {
         self.state.meta(self.index)
     }
@@ -232,6 +238,7 @@ impl<'a> RDesRing<'a> {
         let descriptor = &mut self.descriptors[self.index];
         assert!(descriptor.available());
 
+        #[cfg(feature = "ptp")]
         self.state.clear(self.index);
         self.descriptors[self.index].set_ready(self.buffers[self.index].0.as_mut_ptr());
