+++ embassy-stm32/src/eth/v1/tx_desc.rs
@@ -1,9 +1,11 @@
 use core::sync::atomic::{Ordering, compiler_fence, fence};
 
+#[cfg(feature = "ptp")]
 use embassy_net_driver::PacketMeta;
 use vcell::VolatileCell;
 
 use crate::eth::TX_BUFFER_SIZE;
+#[cfg(feature = "ptp")]
 use crate::eth::packet_state::TxPacketStateRing;
 use crate::pac::ETH;
 
@@ -108,6 +110,7 @@ impl TDes {
 pub(crate) struct TDesRing<'a> {
     descriptors: &'a mut [TDes],
     buffers: &'a mut [Packet<TX_BUFFER_SIZE>],
+    #[cfg(feature = "ptp")]
     state: TxPacketStateRing<'a>,
     index: usize,
 }
@@ -117,7 +120,7 @@ impl<'a> TDesRing<'a> {
     pub(crate) fn new(
         descriptors: &'a mut [TDes],
         buffers: &'a mut [Packet<TX_BUFFER_SIZE>],
-        state: TxPacketStateRing<'a>,
+        #[cfg(feature = "ptp")] state: TxPacketStateRing<'a>,
     ) -> Self {
         assert!(descriptors.len() > 0);
         assert!(descriptors.len() == buffers.len());
@@ -134,6 +137,7 @@ impl<'a> TDesRing<'a> {
         Self {
             descriptors,
             buffers,
+            #[cfg(feature = "ptp")]
             state,
             index: 0,
         }
@@ -142,8 +146,6 @@ impl<'a> TDesRing<'a> {
     pub(crate) fn len(&self) -> usize {
         self.descriptors.len()
     }
-
-    pub(crate) fn collect_completed(&mut self) {}
 
     /// Return the next available packet buffer for transmitting, or None
     pub(crate) fn available(&mut self) -> Option<&mut [u8]> {
@@ -155,6 +157,7 @@ impl<'a> TDesRing<'a> {
         }
     }
 
+    #[cfg(feature = "ptp")]
     pub(crate) fn set_meta(&mut self, meta: PacketMeta) {
         self.state.set_meta(meta);
     }
@@ -166,6 +169,7 @@ impl<'a> TDesRing<'a> {
 
         descriptor.set_buffer1(self.buffers[self.index].0.as_ptr());
         descriptor.set_buffer1_len(len);
+        #[cfg(feature = "ptp")]
         self.state.commit(self.index);
 
         descriptor.set_owned();
