+++ embassy-stm32/src/eth/v2/mod.rs
@@ -512,13 +512,24 @@ impl<'d, T: Instance, P: Phy> Ethernet<'d, T, P> {
             w.set_rbsz(RX_BUFFER_SIZE as u16);
         });
 
+        #[cfg(feature = "ptp")]
         let (tx_state, rx_state) = queue.packet_state.split();
 
         let mut this = Self {
             _peri: peri,
             _wake_guard: T::RCC_INFO.wake_guard(),
-            tx: TDesRing::new(&mut queue.tx_desc, &mut queue.tx_buf, tx_state),
-            rx: RDesRing::new(&mut queue.rx_desc, &mut queue.rx_buf, rx_state),
+            tx: TDesRing::new(
+                &mut queue.tx_desc,
+                &mut queue.tx_buf,
+                #[cfg(feature = "ptp")]
+                tx_state,
+            ),
+            rx: RDesRing::new(
+                &mut queue.rx_desc,
+                &mut queue.rx_buf,
+                #[cfg(feature = "ptp")]
+                rx_state,
+            ),
             _pins: pins,
             phy,
             mac_addr,
